# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: 집별 서빙 API
  version: 0.1.0
  description: >
    비회원 개인화 부동산 레이더 MVP API. enum 코드는 enums.yaml(SSOT)을 따른다.
    모든 시각은 Asia/Seoul. 에러는 공통 envelope(ErrorResponse)을 쓴다.
servers:
  - url: http://localhost:8080
    description: 로컬
...
components:
  securitySchemes:
    anonymousId:
      type: apiKey
      in: header
      name: X-Anonymous-Id
    internalToken:
      type: apiKey
      in: header
      name: X-Internal-Token
  responses:
    Error400:
      description: 검증 실패
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
...
    Error404:
      description: 리소스 없음
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, traceId]
          properties:
            code:
              type: string
              description: "INVALID_REGION | INVALID_DATE_RANGE | INVALID_ENUM | UNAUTHORIZED | ANONYMOUS_USER_NOT_FOUND | DEVICE_NOT_FOUND | ANNOUNCEMENT_NOT_FOUND | ALREADY_EXISTS | RATE_LIMITED | UPSTREAM_UNAVAILABLE | INTERNAL_ERROR"
            message: { type: string }
            traceId: { type: string }
...
    UnitMargin:
      type: object
      properties:
        houseType: { type: [string, 'null'] }
        areaM2: { type: [number, 'null'] }
        supplyAmountManwon: { type: [integer, 'null'], format: int64 }
        marketMedianManwon: { type: [integer, 'null'], format: int64 }
        marginManwon: { type: [integer, 'null'], format: int64 }
        marginRatio: { type: [number, 'null'] }
        grade: { type: string, enum: [HIGH, MID, LOW, UNAVAILABLE] }
        sampleCount: { type: integer, format: int64 }
    AnnouncementMargin:
      type: object
      properties:
        announcementId: { type: integer, format: int64 }
        priceCap: { type: boolean }
        unranked: { type: boolean }
        representativeGrade: { type: string, enum: [HIGH, MID, LOW, UNAVAILABLE] }
        basisRegion: { type: string }
        basisMonths: { type: integer }
        units: { type: array, items: { $ref: '#/components/schemas/UnitMargin' } }
    AnnouncementSummary:
      type: object
      properties:
        id: { type: integer, format: int64 }
        pblancNo: { type: [string, 'null'] }
        title: { type: [string, 'null'] }
        supplyType: { type: [string, 'null'] }
        regionName: { type: [string, 'null'] }
        applyStart: { type: [string, 'null'], format: date }
        applyEnd: { type: [string, 'null'], format: date }
        winnerAnnounceDate: { type: [string, 'null'], format: date }
        sourceName: { type: string }
        sourceUrl: { type: [string, 'null'] }
    AnnouncementList:
      type: object
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/AnnouncementSummary' } }
        page: { type: integer }
        size: { type: integer }
        totalCount: { type: integer, format: int64 }
    AnnouncementDetail:
      allOf:
        - $ref: '#/components/schemas/AnnouncementSummary'
        - type: object
          properties:
            bjdCode: { type: [string, 'null'] }
            contractDate: { type: [string, 'null'], format: date }
            summary: { type: object, additionalProperties: true }
            collectedAt: { type: string, format: date-time }
            updatedAt: { type: string, format: date-time }
            margin: { oneOf: [ { $ref: '#/components/schemas/AnnouncementMargin' }, { type: 'null' } ] }
...
paths:
...
  /v1/announcements:
    get:
      tags: [serving]
      summary: 공고 목록
      parameters:
        - { name: region, in: query, schema: { type: string } }
        - { name: supplyType, in: query, schema: { type: string } }
        - { name: page, in: query, schema: { type: integer, default: 0 } }
        - { name: size, in: query, schema: { type: integer, default: 20, maximum: 100 } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/AnnouncementList' } } } }
        '400': { $ref: '#/components/responses/Error400' }
  /v1/announcements/{id}:
    get:
      tags: [serving]
      summary: 공고 상세
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/AnnouncementDetail' } } } }
        '404': { $ref: '#/components/responses/Error404' }
...